changed the routing for searches so they do not conflict with getting tasks
